fix(tool-registry): break index.js ⇄ ensure.js import cycle — singleton moves to default-registry.ts
Biome noImportCycles (Tier A CI gate) counts dynamic imports too, so the
lazy import() could not save it. getDefaultRegistry/peekGlobalRegistry/
_resetDefaultRegistry now live in default-registry.ts (own module, no
back-edge); index.ts re-exports them unchanged for all consumers; ensure.ts
imports the singleton module directly. globalThis symbol key preserved.

See change: add-skill-tool-provisioning
