+++ packages/shared/src/tool-registry/AGENTS.md
@@ -4,7 +4,8 @@ Files in this directory. One row per source file.
 
 | File | Purpose |
 |------|---------|
-| `definitions.ts` | Registers standard tool set. Each definition declares ordered strategy chain (override → bare-import → managed → npm-global → where) + classifier (strategy → source). Includes build-time tools `electron` + `node-pty` (registered as `kind: "module"` returning package directory; resolved hoist-aware via `bareImportPackageDirStrategy` with optional `searchPaths`). See change: register-build-time-tools. Registers `binaryDef("bash")` on every platform. `INSTALL_HINTS` map (bash, git, node, gh, zrok [v2: doc URLs `docs.zrok.io/getting-started`, win32 `zrok2` on PATH; see change: support-zrok-v2], ngrok, tailscale, zerotier-cli) attached via `binaryDef`; registers `ngrok`/`tailscale`/`zerotier-cli` binaries for the tunnel providers (See change: add-tunnel-providers). `npx` unchanged, no installHints. See change: register-bash-and-tool-install-help. bash chain now override → bundled-git-bash → managed → where (win32 surfaces bundled sh.exe before PATH). classify maps `bundled-git-bash` → source `bundled`. See change: resolve-bundled-bash-on-windows. wmic no longer registered on win32 (removed by default Win 11 22H2+); win32 set = powershell, tasklist, taskkill. See change: replace-wmic-with-powershell. Media/skill-facing tools: ffmpeg (override → static-npm(ffmpeg-static) → where), ffprobe (static-npm @ffprobe-installer/ffprobe), imagemagick (where `convert`), chromium (pw-browser), agent-browser (where; hint `pi install npm:pi-agent-browser`), pi-doc-engine (docker-image; hint `npm run build:image` requiresConfirm) — installHints with docsAnchors install-ffmpeg/install-ffprobe/install-imagemagick/install-chromium. See change: add-skill-tool-provisioning. |
+| `definitions.ts` | Registers standard tool set. Each definition declares ordered strategy chain (override → bare-import → managed → npm-global → where) + classifier (strategy → source). Includes build-time tools `electron` + `node-pty` (registered as `kind: "module"` returning package directory; resolved hoist-aware via `bareImportPackageDirStrategy` with optional `searchPaths`). See change: register-build-time-tools. Registers `binaryDef("bash")` on every platform. `INSTALL_HINTS` map (bash, git, node, gh, zrok [v2: doc URLs `docs.zrok.io/getting-started`, win32 `zrok2` on PATH; see change: support-zrok-v2], ngrok, tailscale, zerotier-cli) attached via `binaryDef`; registers `ngrok`/`tailscale`/`zerotier-cli` binaries for the tunnel providers (See change: add-tunnel-providers). `npx` unchanged, no installHints. See change: register-bash-and-tool-install-help. bash chain now override → bundled-git-bash → managed → where (win32 surfaces bundled sh.exe before PATH). classify maps `bundled-git-bash` → source `bundled`. See change: resolve-bundled-bash-on-windows. wmic no longer registered on win32 (removed by default Win 11 22H2+); win32 set = powershell, tasklist, taskkill. See change: replace-wmic-with-powershell. Media/skill-facing tools: ffmpeg (override → static-npm(ffmpeg-static) → where), ffprobe (static-npm @ffprobe-installer/ffprobe), imagemagick (where `convert`; win32 platformStrategies probe `magick` — System32 convert.exe is the NTFS tool), chromium (pw-browser), agent-browser (where; hint `pi install npm:pi-agent-browser`), pi-doc-engine (docker-image; hint `npm run build:image` requiresConfirm) — installHints with docsAnchors install-ffmpeg/install-ffprobe/install-imagemagick/install-chromium. See change: add-skill-tool-provisioning. |
+| `default-registry.ts` | Process-wide `ToolRegistry` singleton. Own module (NOT index.ts) so ensure.ts can import getDefaultRegistry without an index.js round trip (Biome noImportCycles). Exports `getDefaultRegistry` (lazy, publishes on `globalThis[Symbol.for("pi-dashboard.tool-registry")]`), `peekGlobalRegistry` (cycle-safe accessor for `platform/runner.ts`), `_resetDefaultRegistry` (test-only). See change: add-skill-tool-provisioning. |
 | `ensure-cli.ts` | `pi-dashboard-ensure` CLI (design D5). Reads a package's `pi.tools`, ingests, ensures, prints report. Flags `--install`, `--json` (--json always exits 0); exit 0 when all required present/installed. TTY confirm prompt for requiresConfirm; headless auto-deny. Run via tsx launcher `packages/shared/bin/pi-dashboard-ensure.mjs`. See change: add-skill-tool-provisioning. |
 | `ensure.ts` | `ensureTools` — library face of tool ensuring. Recommend-only default; `autoInstall:true` runs ONLY resolved first-party `installHints.commands[pkgmgr]`; `requiresConfirm` hints need per-invocation confirm (no callback = auto-deny). Actions present/recommended/installed/degraded/blocked; `ok:false` ⇔ required tool not present/installed. Never throws for missing tools. See change: add-skill-tool-provisioning. |
 | `index.ts` | Tool registry public entry. Re-exports `types`, `OverridesStore`/`defaultOverridesPath`, `ToolRegistry`, `registerDefaultTools`, `strategies`, `parseSkillTools`/`ingestSkillTools`/`discoverSkillManifests`/`ingestInstalledSkillTools`, `ensureTools`. Exports `getDefaultRegistry` (lazy process-wide singleton, publishes on `globalThis[Symbol.for("pi-dashboard.tool-registry")]`), `peekGlobalRegistry` (cycle-safe accessor for `platform/runner.ts`), `_resetDefaultRegistry` (test-only). |
